refactor: products -> 전역 상태 변경
