+++ src/App.js
@@ -1,8 +1,6 @@
-import { useEffect, useState } from 'react';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import axios from 'axios';
 import {
   NavBar,
   ShoppingCart,
@@ -17,41 +15,15 @@ import rootReducer from './reducers';
 const store = createStore(rootReducer);
 
 function App() {
-  const [products, setProducts] = useState([]);
-
-  useEffect(() => {
-    // TODO: 미들 웨어 적용
-    async function fetchProducts() {
-      try {
-        const response = await axios.get('/products');
-
-        setProducts(response.data);
-      } catch (error) {
-        //TODO: 상품을 못 받아 왔을 때, 안내 화면 띄우기
-        console.error(error.message);
-      }
-    }
-
-    fetchProducts();
-  }, []);
-
   return (
     <Provider store={store}>
       <GlobalStyle />
       <Router>
         <NavBar />
-        <Route exact path={[ROUTE.HOME, ROUTE.PRODUCTS]}>
-          <Products products={products} />
-        </Route>
-        <Route exact path={ROUTE.CART}>
-          <ShoppingCart />
-        </Route>
-        <Route exact path={ROUTE.ORDER_PAYMENT}>
-          <OrderPayment />
-        </Route>
-        <Route exact path={ROUTE.COMPLETED_ORDER}>
-          <CompletedOrder />
-        </Route>
+        <Route exact path={[ROUTE.HOME, ROUTE.PRODUCTS]} component={Products} />
+        <Route exact path={ROUTE.CART} component={ShoppingCart} />
+        <Route exact path={ROUTE.ORDER_PAYMENT} component={OrderPayment} />
+        <Route exact path={ROUTE.COMPLETED_ORDER} component={CompletedOrder} />
       </Router>
     </Provider>
   );
