+++ src/components/pages/CompletedOrder/index.styles.js
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
 
 export const Main = styled.main`
+  height: 100%;
   background-color: var(--color-background);
   padding: 7rem 15rem;
